Implement unobserving a single product via it's product card - account's tab section also has dedicated buttons to unobserve each product; Move product observability to a separate component and improve it's integration with storage service;
